Implemented some of the easier operations.

+++ src/chip8.h
@@ -12,7 +12,7 @@ class CHIP8 {
 public:
   std::uint16_t opcode;
   std::array<std::uint8_t, 16> V; // 16 8-bit data registers
-  std::uint8_t pc; // program counter
+  std::uint16_t pc; // program counter
   std::uint16_t I; // 16-bit index register
   std::array<std::uint8_t, 4096> mem; // 4096 bytes of addressable memory
   std::uint8_t stack_pointer; // 8-bit stack pointer
